backend/requirements.txt
fastapi
pymongo
uvicorn
python-multipart
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]